Extract fetch, fix data bug Contacts only used in thread component. Thus more sensible to be part of that component. API returns users as users: [], not []. Access accordingly. Set text area value to blank string to silence warning.
