+++ src/components/Messages/Messages.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import styles from './Messages.module.css'; // Import CSS Module
+import styles from './Messages.module.css'; 
 import Sidebar from '../Sidebar/Sidebar';
 import Thread from '../Thread/Thread';
 
@@ -7,7 +7,6 @@ import Thread from '../Thread/Thread';
 function Messages() {
   const [selectedThread, setSelectedThread] = useState(false);
   const [newThreadForm, setNewThreadForm] = useState(false)
-  const [contacts, setContacts] = useState(null);
   const dummyThread = ['hey', 'hi', 'hoo'];
   const handleThreadSelect = (thread) => {
     setSelectedThread(thread);
@@ -15,24 +14,6 @@ function Messages() {
   const handleNewThreadClick = () => {
     setNewThreadForm(true);
   }
-  useEffect(()=> {
-    const fetchContacts = async() => {
-      const token = Cookies.get('jwt_token');
-      const response = await fetch(`http://localhost:3000/users`, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: 'Bearer ' + token
-        }
-      });
-      if (!response.ok) {
-        console.error('Error fetching contacts');
-      }
-      const data = await response.json();
-      console.log(data)
-      setContacts(data);}
-    fetchContacts()
-  },[])
 
   return (
     <div className={styles.messagesContainer}>
@@ -43,7 +24,7 @@ function Messages() {
         
 
       />
-      <Thread newThreadForm={newThreadForm} newThreadContacts={[{username:'j'}, {username: 'k'}]} selectedThread={selectedThread} />
+      <Thread newThreadForm={newThreadForm}  selectedThread={selectedThread} />
     </div>
   );
 }
